Fixing bug in vignette
When I originally made changes to the vignette to make it work without the cgdsr package, I introduced a bug where not all the required data (i.e. one table per gene) was used. With this commit I aim to fix this by uploading all the required tables and changing the code in the vignette.